feat: move to separate file and remove tls toggle for dex server
