+++ config/advanced-install/namespaced-numaflow-server.yaml
@@ -247,12 +247,6 @@ spec:
               key: server.address
               name: numaflow-cmd-params-config
               optional: true
-        - name: NUMAFLOW_DISABLE_DEX_SERVER_TLS
-          valueFrom:
-            configMapKeyRef:
-              key: server.dex.disable.tls
-              name: numaflow-cmd-params-config
-              optional: true
         image: quay.io/numaproj/numaflow:latest
         imagePullPolicy: Always
         livenessProbe:
