+++ config/advanced-install/numaflow-server.yaml
@@ -258,12 +258,6 @@ spec:
               key: server.address
               name: numaflow-cmd-params-config
               optional: true
-        - name: NUMAFLOW_DISABLE_DEX_SERVER_TLS
-          valueFrom:
-            configMapKeyRef:
-              key: server.dex.disable.tls
-              name: numaflow-cmd-params-config
-              optional: true
         image: quay.io/numaproj/numaflow:latest
         imagePullPolicy: Always
         livenessProbe:
