+++ config/base/numaflow-server/numaflow-server-deployment.yaml
@@ -120,12 +120,6 @@ spec:
                 name: numaflow-cmd-params-config
                 key: server.address
                 optional: true
-          - name: NUMAFLOW_DISABLE_DEX_SERVER_TLS
-            valueFrom:
-              configMapKeyRef:
-                name: numaflow-cmd-params-config
-                key: server.dex.disable.tls
-                optional: true
           resources:
             limits:
               cpu: 500m
