+++ config/namespace-install.yaml
@@ -16572,7 +16572,7 @@ spec:
         - containerPort: 5556
         volumeMounts:
         - mountPath: /etc/numaflow/dex/cfg/config.yaml
-          name: new-config
+          name: generated-dex-config
           subPath: config.yaml
         - mountPath: /etc/numaflow/dex/tls/tls.crt
           name: tls
@@ -16601,11 +16601,11 @@ spec:
         name: dex-init
         volumeMounts:
         - mountPath: /cfg
-          name: old-config
+          name: connector-config
         - mountPath: /tls
           name: tls
         - mountPath: /tmp
-          name: new-config
+          name: generated-dex-config
       serviceAccountName: numaflow-dex-server
       volumes:
       - configMap:
@@ -16613,11 +16613,11 @@ spec:
           - key: config.yaml
             path: config.yaml
           name: numaflow-dex-server-config
-        name: old-config
+        name: connector-config
       - emptyDir: {}
         name: tls
       - emptyDir: {}
-        name: new-config
+        name: generated-dex-config
 ---
 apiVersion: apps/v1
 kind: Deployment
@@ -16691,12 +16691,6 @@ spec:
           valueFrom:
             configMapKeyRef:
               key: server.address
-              name: numaflow-cmd-params-config
-              optional: true
-        - name: NUMAFLOW_DISABLE_DEX_SERVER_TLS
-          valueFrom:
-            configMapKeyRef:
-              key: server.dex.disable.tls
               name: numaflow-cmd-params-config
               optional: true
         image: quay.io/numaproj/numaflow:latest
